requirements.txt
asgiref==3.7.2
awscli==1.31.4
boto3==1.33.4
botocore==1.33.4
colorama==0.4.4
Django==4.2.7
django-cors-headers==4.3.1
django-jazzmin==2.6.0
docutils==0.16
gunicorn==21.2.0
jmespath==1.0.1
packaging==23.2
pyasn1==0.5.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
PyYAML==6.0.1
rsa==4.7.2
s3transfer==0.8.2
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.7